ceate get and update url repo

+++ app/migrations/20210720171902_initial_urls_table.up.sql
@@ -3,7 +3,7 @@ CREATE TABLE urls (
       short_code varchar(15) unique,
       full_url text,
       expiry timestamptz,
-      hits numeric(36) default 0,
+      hits int default 0,
       created_at timestamptz,
       updated_at timestamptz,
       deleted_at timestamptz
